added a function in simple-grading.js
if pressed left arrow key, the container will scroll left. To be honest, I don't understand why this code is working.  In this line,                 container.scrollLeft="0px";
i set scrollLeft to 0px, I've tested it with 30, 300. they all work the same.
So, I'm guessing every time left arrow key is pressed the container will scroll left and re-focus on the input box?
I think this is not a really elegant solution, but could be a temporary solution, I'm new to css and javascript and can't really come up with a better solution.